Update lib/bindings.js
AppJS only supports 64 bit node on Linux. The only time appjs.node will not be found is if the platform is unsupported or if the user has changed/deleted files they should not have.

There is no need to have multiple checks for which platform and arch we're dealing with. Either appjs.node is found or it isn't, that's taken care of by having the directory structure of the bindings reflect the platform and architecture.

Note that I'm using Windows 7 x64 but the platform still says win32. It's a generic name for windows platform, the arch is accurate.